juggler: refactor filenames
