+++ senders/juggler/juggler.go
@@ -2,23 +2,14 @@ package juggler
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
-	"net/http"
-	"os"
 	"path"
-	"reflect"
 	"sync"
 	"sync/atomic"
 
-	"github.com/combaine/combaine/common/configs"
-	"github.com/combaine/combaine/common/httpclient"
 	"github.com/combaine/combaine/common/logger"
 	"github.com/combaine/combaine/common/tasks"
 	lua "github.com/yuin/gopher-lua"
-	yaml "gopkg.in/yaml.v2"
 )
 
 const (
@@ -27,263 +18,18 @@ const (
 	DEFAULT_CHECK_LEVEL = "OK"
 )
 
-type Conditions struct {
-	OK   []string `codec:"OK"`
-	INFO []string `codec:"INFO"`
-	WARN []string `codec:"WARN"`
-	CRIT []string `codec:"CRIT"`
-}
-
-type JugglerConfig struct {
-	PluginsDir         string                  `codec:"PluginsDir"`
-	Plugin             string                  `codec:"Plugin"`
-	DefaultCheckStatus string                  `codec:"DefaultCheckStatus"`
-	Host               string                  `codec:"Host"`
-	Methods            []string                `codec:"Methods"`
-	Aggregator         string                  `codec:"Aggregator"`
-	CheckName          string                  `codec:"checkname"`
-	Description        string                  `codec:"description"`
-	AggregatorKWargs   JugglerAggregatorKWArgs `codec:"aggregator_kwargs"`
-	Flap               JugglerFlapConfig       `codec:"flap"`
-	JPluginConfig      configs.PluginConfig    `codec:"config"`
-	JHosts             []string                `codec:"juggler_hosts"`
-	JFrontend          []string                `codec:"juggler_frontend"`
-	Conditions
-}
-
-type jugglerSenderConf struct {
-	PluginsDir string   `yaml:"PluginsDir"`
-	Hosts      []string `yaml:"juggler_hosts"`
-	Frontend   []string `yaml:"juggler_frontend"`
-}
-
 type jugglerSender struct {
-	JugglerConfig
+	*JugglerConfig
 	id    string
 	state *lua.LState
 }
 
-// GetJugglerConfig read yaml file with two arrays of hosts
-// if juggler_frontend not defined, use juggler_hosts as frontend
-func GetJugglerConfig() (conf jugglerSenderConf, err error) {
-	var path string = os.Getenv("JUGGLER_CONFIG")
-	if len(path) == 0 {
-		path = DEFAULT_CONFIG_PATH
-	}
-
-	rawConfig, err := ioutil.ReadFile(path)
-	if err != nil {
-		return
-	}
-	err = yaml.Unmarshal(rawConfig, &conf)
-	if conf.Frontend == nil {
-		conf.Frontend = conf.Hosts
-	}
-	if conf.PluginsDir == "" {
-		conf.PluginsDir = DEFAULT_PLUGIN_DIR
-	}
-	return
-}
-
-func NewJugglerSender(conf JugglerConfig, id string) (*jugglerSender, error) {
+func NewJugglerSender(conf *JugglerConfig, id string) (*jugglerSender, error) {
 	return &jugglerSender{
 		JugglerConfig: conf,
 		id:            id,
 		state:         nil,
 	}, nil
-}
-
-// LoadPlugin cleanup lua state global/local environment
-// and load lua plugin by name from juggler config section
-func LoadPlugin(fileName string) (*lua.LState, error) {
-	l := lua.NewState()
-	if err := l.DoFile(fileName); err != nil {
-		return nil, err
-	}
-	// TODO: overwrite/cleanup globals in lua plugin?
-	//		 cache lua state in plugin cache?
-	return l, nil
-}
-
-// preparePluginEnv add data from aggregate task as global variable in lua
-// plugin. Also inject juggler conditions from juggler configs and plugin config
-func (js *jugglerSender) preparePluginEnv(taskData tasks.DataType) error {
-	ltable, err := dataToLuaTable(js.state, taskData)
-	if err != nil {
-		return fmt.Errorf("Failed to convert taskData to lua table: %s", err)
-	}
-
-	js.state.SetGlobal("payload", ltable)
-
-	levels := make(map[string][]string)
-	if js.Conditions.OK != nil {
-		levels["OK"] = js.Conditions.OK
-	}
-	if js.Conditions.INFO != nil {
-		levels["INFO"] = js.Conditions.INFO
-	}
-	if js.Conditions.WARN != nil {
-		levels["WARN"] = js.Conditions.WARN
-	}
-	if js.Conditions.CRIT != nil {
-		levels["CRIT"] = js.Conditions.CRIT
-	}
-
-	lconditions := js.state.NewTable()
-	idx := 0
-	for name, cond := range levels {
-		lcondTable := js.state.NewTable()
-		lconditions.RawSetString(name, lua.LNumber(idx))
-		lconditions.RawSetInt(idx, lcondTable)
-		for _, v := range cond {
-			lcondTable.Append(lua.LString(v))
-		}
-		idx++
-	}
-	js.state.SetGlobal("conditions", lconditions)
-
-	if lconfig, err := jPluginConfigToLuaTable(js.state, js.JPluginConfig); err != nil {
-		return err
-	} else {
-		js.state.SetGlobal("config", lconfig)
-	}
-	return nil
-}
-
-// runPlugin run lua plugin with prepared environment
-// collect, convert and return plugin result
-func (js *jugglerSender) runPlugin() ([]jugglerEvent, error) {
-	js.state.Push(js.state.GetGlobal("run"))
-	if err := js.state.PCall(0, 1, nil); err != nil {
-		return nil, err
-	}
-	result := js.state.ToTable(1)
-	events, err := luaResultToJugglerEvents(js.DefaultCheckStatus, result)
-	if err != nil {
-		return nil, err
-	}
-	return events, nil
-}
-
-// getCheck query juggler api for check and Unmarshal json response in to
-// JugglerResponse type
-func (js *jugglerSender) getCheck(ctx context.Context) (JugglerResponse, error) {
-	var hostChecks JugglerResponse
-	var flap map[string]map[string]JugglerFlapConfig
-
-	for jhost := range js.JHosts {
-		url := fmt.Sprintf(getCheckUrl, jhost, js.Host)
-		logger.Infof("%s Query check %s", js.id, url)
-
-		resp, err := httpclient.Get(ctx, url)
-		switch err {
-		case nil:
-			body, rerr := ioutil.ReadAll(resp.Body)
-			if rerr != nil {
-				logger.Errf("%s %s", js.id, rerr)
-				continue
-			}
-			if resp.StatusCode != http.StatusOK {
-				return nil, errors.New(string(body))
-			}
-			if err := json.Unmarshal(body, &hostChecks); err != nil {
-				return nil, err
-			}
-			if err := json.Unmarshal(body, &flap); err != nil {
-				return nil, err
-			}
-			for c, v := range flap[js.Host] {
-				chk := hostChecks[js.Host][c]
-				chk.Flap = v
-				hostChecks[js.Host][c] = chk
-			}
-			return hostChecks, nil
-		case context.Canceled, context.DeadlineExceeded:
-			return nil, err
-		default:
-			logger.Errf("%s %s", js.id, err)
-		}
-	}
-
-	return nil, errors.New("Failed to get juggler check")
-
-}
-func (js *jugglerSender) ensureFlap(jcheck *JugglerCheck) error {
-	if js.JugglerConfig.Flap.Enable == 1 {
-		jcheck.Flap.Enable = 1
-		if jcheck.Flap != js.JugglerConfig.Flap {
-			jcheck.Flap = js.JugglerConfig.Flap
-			jcheck.Update = true
-		}
-	} else {
-		jcheck.Flap = JugglerFlapConfig{}
-	}
-	return nil
-}
-
-// ensureCheck check that juggler check exists and it in sync with task data
-// if need it create or update check
-func (js *jugglerSender) ensureCheck(ctx context.Context, triggers []jugglerEvent) error {
-	hostChecks, err := js.getCheck(ctx)
-	if err != nil {
-		return err
-	}
-
-	services, ok := hostChecks[js.Host]
-	if !ok {
-		services = make(map[string]JugglerCheck)
-		hostChecks[js.Host] = services
-	}
-	childSet := make(map[string]struct{}) // set
-	for n, v := range services {
-		for _, c := range v.Children {
-			childSet[c.Host+":"+n] = struct{}{}
-		}
-	}
-
-	for _, t := range triggers {
-		check, ok := services[t.Service]
-		if !ok {
-			check = JugglerCheck{Update: true}
-		}
-		if t.Host == js.Host { // for metahost
-			if err := js.ensureFlap(&check); err != nil {
-				return err
-			}
-			if check.Aggregator != js.Aggregator ||
-				!reflect.DeepEqual(check.AggregatorKWArgs, js.AggregatorKWargs) {
-
-				check.Update = true
-				check.Aggregator = js.Aggregator
-				check.AggregatorKWArgs = js.AggregatorKWargs
-			}
-
-		} else {
-			if _, ok := childSet[t.Host+":"+t.Service]; !ok {
-				check.Update = true
-				check.Children = append(check.Children, JugglerChildrenCheck{
-					Host:    t.Host,
-					Service: t.Service,
-				})
-			}
-		}
-		if check.Update {
-			if err := js.updateCheck(ctx, check); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func (js *jugglerSender) updateCheck(ctx context.Context, check JugglerCheck) error {
-	return nil
-}
-
-// sendEvent send juggler event borned by ensureCheck to juggler's
-func (js *jugglerSender) sendEvent(event jugglerEvent) error {
-	return nil
 }
 
 // Send make all things abount juggler sender tasks
